// frontend/src/features/parser/ParseText/model/parse-text.types.ts
export interface ParseTextState {
	isLoading: boolean
	isLoaded: boolean
	isFailed: boolean
	error: null | string
}

export enum ParseTextActionsTokens {
	LOADING_PARSE_TEXT_INFO_ACTION = 'LOADING_PARSE_TEXT_INFO_ACTION',
	LOADED_PARSE_TEXT_INFO_ACTION = 'LOADED_PARSE_TEXT_INFO_ACTION',
	ERROR_IN_LOADING_PARSE_TEXT_INFO_ACTION = 'ERROR_IN_LOADING_PARSE_TEXT_INFO_ACTION',

	CLEAR_STATE_PARSE_TEXT_ACTION = 'CLEAR_STATE_PARSE_TEXT_ACTION',

	ASYNC_LOAD_PARSE_TEXT_INFO_ACTION = 'ASYNC_LOAD_PARSE_TEXT_INFO_ACTION'
}